Added casting uint value to long in order to make 32 bit shifting correct (#33)
* Added casting uint value to long in order to make 32 bit shifting correct

* Added unit test for PacLogonInfo

+++ Kerberos.NET/Entities/Authorization/NdrBinaryReader.cs
@@ -96,7 +96,7 @@ namespace Kerberos.NET.Entities.Authorization
 
             if (low != 0xffffffffL && high != 0x7fffffffL)
             {
-                var fileTime = (high << 32) + low;
+                var fileTime = ((long)high << 32) + low;
 
                 var universalTicks = fileTime + FileTimeOffset;
